requirements.txt
Flask==3.1.0
Flask-Limiter==3.13
Flask-WTF==1.2.2
python-dotenv==1.0.0
gunicorn==23.0.0
